allow to pass a function in blacklist option
this function accepts the package config of the specific version as argument and can decide according to information inside that configuration whether this module can be provided by unpkg. this can be very useful for people who want to expose only client-side javascript modules from their private npm registry, but not server-side node modules. strategy of how to distinguish client from server side is of course up to the user and handled by the blacklist function.